fixed size of push button
