+++ pro_110822/client_view/clientviewupperframe.py
@@ -1,10 +1,6 @@
-#pro_110822/firstOpenView.py
-
 from PySide6 import QtCore, QtWidgets
 from PySide6.QtCore import QRect
-
 from PySide6.QtGui import QFont
-
 from PySide6.QtWidgets import (
     QApplication,
     QMainWindow,
@@ -18,11 +14,13 @@ from PySide6.QtWidgets import (
     QFrame,
     QSizePolicy
 )
+from . import clientviewupperframebtn1
 
 
 class ClientViewUpperFrame(QtWidgets.QFrame):
     def __init__(self, *args, **kwargs):
         super(ClientViewUpperFrame, self).__init__(*args, **kwargs)
+        self.searchBtton : QtWidgets.QPushButton
 
         self._set_style()
         self._initUi()
@@ -34,69 +32,48 @@ class ClientViewUpperFrame(QtWidgets.QFrame):
                             "    color: #fff;")
         self.setFrameShape(QFrame.StyledPanel)
         self.setFrameShadow(QFrame.Raised)
-        # self.setContentsMargins(0,0,0,0)
+
 
 
     def _initUi(self):
-        #Search for servers button
-        self.searchButton = QPushButton(self)
-        self.searchButton.setObjectName(u"pushButton")
-        self.searchButton.setGeometry(QRect(450, 70, 131, 31))
-        sizePolicy = QSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.searchButton.sizePolicy().hasHeightForWidth())
-        self.searchButton.setSizePolicy(sizePolicy)
-        font = QFont()
-        font.setBold(True)
-        self.searchButton.setFont(font)
-        self.searchButton.setStyleSheet(u"QPushButton::flat\n"
-                                        "{\n"
-                                        "       background-color: transparent;\n"
-                                        "       border: none;\n"
-                                        "       color: #fff;\n"
-                                        "}\n"
-                                        "\n"
-                                        "QPushButton::hover\n"
-                                        "{\n"
-                                        "       background-color: #04c91b;\n"
-                                        "       border: 1px solid #04c91b;\n"
-                                        "}\n"
-                                        "\n"
-                                        "QPushButton::pressed\n"
-                                        "{\n"
-                                        "       background-color: #02f01e;\n"
-                                        "       border: 1px solid #02f01e;\n"
-                                        "}\n"
-                                        "")
-        self.searchButton.setFlat(False)
+        self.searchBtton = clientviewupperframebtn1.ClientViewUpperFrameBtn1(self)
+        self.searchBtton.set_size(100, 199)
+        self.searchBtton.set_text('Search for servers')
+        # self.searchBtton.setText('Search for servers')
+        # self.searchBtton.set_geometry(450, 70, 131, 131)
 
-        #Create a server button
-        self.createButton = QPushButton(self)
-        self.createButton.setObjectName(u"pushButton_2")
-        self.createButton.setGeometry(QRect(300, 70, 131, 31))
-        sizePolicy.setHeightForWidth(self.createButton.sizePolicy().hasHeightForWidth())
-        self.createButton.setSizePolicy(sizePolicy)
-        self.createButton.setFont(font)
-        self.createButton.setStyleSheet(u"QPushButton::flat\n"
-                                        "{\n"
-                                        "       background-color: transparent;\n"
-                                        "       border: none;\n"
-                                        "       color: #fff;\n"
-                                        "}\n"
-                                        "\n"
-                                        "QPushButton::hover\n"
-                                        "{\n"
-                                        "       background-color: #0279f0;\n"
-                                        "       border: 1px solid #0279f0;\n"
-                                        "}\n"
-                                        "\n"
-                                        "QPushButton::pressed\n"
-                                        "{\n"
-                                        "       background-color: #2590fa;\n"
-                                        "       border: 1px solid #2590fa;\n"
-                                        "}\n"
-                                        "")
+
+
+        # , 450, 70, 131, 131, 'Search for servers'
+
+
+        # #Create a server button
+        # self.createButton = QPushButton(self)
+        # self.createButton.setObjectName(u"pushButton_2")
+        # self.createButton.setGeometry(QRect(300, 70, 131, 31))
+        # sizePolicy.setHeightForWidth(self.createButton.sizePolicy().hasHeightForWidth())
+        # self.createButton.setSizePolicy(sizePolicy)
+        # self.createButton.setFont(font)
+        # self.createButton.setStyleSheet(u"QPushButton::flat\n"
+        #                                 "{\n"
+        #                                 "       background-color: transparent;\n"
+        #                                 "       border: none;\n"
+        #                                 "       color: #fff;\n"
+        #                                 "}\n"
+        #                                 "\n"
+        #                                 "QPushButton::hover\n"
+        #                                 "{\n"
+        #                                 "       background-color: #0279f0;\n"
+        #                                 "       border: 1px solid #0279f0;\n"
+        #                                 "}\n"
+        #                                 "\n"
+        #                                 "QPushButton::pressed\n"
+        #                                 "{\n"
+        #                                 "       background-color: #2590fa;\n"
+        #                                 "       border: 1px solid #2590fa;\n"
+        #                                 "}\n"
+        #                                 "")
         
-        self.searchButton.setText('Search for servers')
-        self.createButton.setText('Create a server')
+        # self.searchButton.setText('Search for servers')
+        # self.createButton.setText('Create a server')
+        pass
